+++ android/src/main/java/com/dieam/reactnativepushnotification/modules/RNPushNotificationConfig.java
@@ -10,6 +10,7 @@ import android.util.Log;
 class RNPushNotificationConfig {
     private static final String KEY_CHANNEL_NAME = "com.dieam.reactnativepushnotification.notification_channel_name";
     private static final String KEY_CHANNEL_DESCRIPTION = "com.dieam.reactnativepushnotification.notification_channel_description";
+    private static final String KEY_NOTIFICATION_FOREGROUND = "com.dieam.reactnativepushnotification.notification_foreground";
     private static final String KEY_NOTIFICATION_COLOR = "com.dieam.reactnativepushnotification.notification_color";
 
     private static Bundle metadata;
@@ -53,6 +54,7 @@ class RNPushNotificationConfig {
         // Default
         return "";
     }
+
     public int getNotificationColor() {
         try {
             int resourceId = metadata.getInt(KEY_NOTIFICATION_COLOR);
@@ -62,5 +64,15 @@ class RNPushNotificationConfig {
         }
         // Default
         return -1;
+    }
+
+    public boolean getNotificationForeground() {
+        try {
+            return metadata.getBoolean(KEY_NOTIFICATION_FOREGROUND, false);
+        } catch (Exception e) {
+            Log.w(RNPushNotification.LOG_TAG, "Unable to find " + KEY_NOTIFICATION_FOREGROUND + " in manifest. Falling back to default");
+        }
+        // Default
+        return false;
     }
 }